prow/artifactuploader: rename from artifact-uploader, fix golint errors
