+++ prow/cmd/artifactuploader/main.go
@@ -26,10 +26,10 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/test-infra/prow/kube"
 
-	"k8s.io/test-infra/prow/artifact-uploader"
+	"k8s.io/test-infra/prow/artifactuploader"
 	"k8s.io/test-infra/prow/gcsupload"
+	"k8s.io/test-infra/prow/kube"
 	"k8s.io/test-infra/prow/logrusutil"
 	"k8s.io/test-infra/prow/pod-utils/options"
 )
@@ -140,7 +140,7 @@ func (o *Options) Run() error {
 		return err
 	}
 
-	controller := artifact_uploader.NewController(client.CoreV1(), prowJobClient, o.Options)
+	controller := artifactuploader.NewController(client.CoreV1(), prowJobClient, o.Options)
 
 	stop := make(chan struct{})
 	defer close(stop)
@@ -161,7 +161,7 @@ func main() {
 	}
 
 	logrus.SetFormatter(
-		logrusutil.NewDefaultFieldsFormatter(nil, logrus.Fields{"component": "artifact-uploader"}),
+		logrusutil.NewDefaultFieldsFormatter(nil, logrus.Fields{"component": "artifactuploader"}),
 	)
 
 	if err := o.Run(); err != nil {
