+++ prow/cmd/deck/static/labels.css
@@ -93,7 +93,7 @@
     color: #ffffff;
 }
 
-.areax00002fprowx00002fartifactx00002duploader {
+.areax00002fprowx00002fartifactuploader {
     background-color: #0052cc;
     color: #ffffff;
 }
